requirements.txt
aiofiles==22.1.0
aiogram==3.0.0b5
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
magic-filter==1.0.9
multidict==6.0.2
pyaes==1.6.1
pydantic==1.10.2
Pyrogram==2.0.59
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.12
yarl==1.8.1
